# src/main/resources/swagger/authApi.yaml
openapi: 3.0.3
info:
  title: Auth Api
  version: 1.0.0

servers:
  - url: /api

paths:
  /auth/register:
    post:
      summary: Register a new user and credential
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/login:
    post:
      summary: Log in with a username and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access and refresh token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/logout:
    post:
      summary: Revoke a refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Logged out

  /auth/me:
    get:
      summary: Get the currently authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'

  /auth/password-reset/request:
    post:
      summary: Email a password reset code to the user with this email address
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: Accepted. Returned whether or not the email belongs to a user.

  /auth/password-reset/confirm:
    post:
      summary: Set a new password using an emailed reset code
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '204':
          description: Password changed

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - name
        - surname
        - email
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        cell:
          type: string

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string

    PasswordResetConfirmRequest:
      type: object
      required:
        - email
        - code
        - newPassword
      properties:
        email:
          type: string
        code:
          type: string
        newPassword:
          type: string
          minLength: 6

    AuthResponse:
      type: object
      required:
        - token
        - refreshToken
      properties:
        token:
          type: string
        refreshToken:
          type: string

    CurrentUserResponse:
      type: object
      required:
        - id
        - username
        - name
        - surname
        - email
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        cell:
          type: string